Space_To_Summer_Sea-muench_destriping/python_implementation/requirements.txt
rasterio==1.4.3
pyyaml==6.0.2
pathlib==1.0.1
numpy==2.3.0
netCDF4==1.7.2
gdal==3.11.0
scipy==1.16.0
scikit-image==0.25.2
PyWavelets==1.8.0
pandas==2.3.1
pyproj==3.7.1
matplotlib==3.10.0
scikit-learn==1.7.1
opencv-python==4.12.0.88
